+++ GeosCore/tracer_mod.F90
@@ -66,6 +66,7 @@ CONTAINS
     USE State_Grid_Mod,   ONLY : GrdState
     USE State_Met_Mod,    ONLY : MetState
     USE Time_Mod,         ONLY : Get_Ts_Chem
+    USE Timers_Mod,       ONLY : Timer_End, Timer_Start
     USE UnitConv_Mod
 
 #if defined( MODEL_GEOS ) || defined( MODEL_GCHP )
@@ -146,6 +147,13 @@ CONTAINS
     !=======================================================================
     ! Convert species units to v/v dry
     !=======================================================================
+
+    ! Halt "All chem" timer (so that unit conv can be timed separately)
+    IF ( Input_Opt%useTimers ) THEN
+       CALL Timer_End( "All chemistry", RC )
+    ENDIF
+
+    ! Convert units
     CALL Convert_Spc_Units(                                                  &
          Input_Opt      = Input_Opt,                                         &
          State_Chm      = State_Chm,                                         &
@@ -160,6 +168,11 @@ CONTAINS
        ErrMsg = 'Unit conversion error (kg -> v/v dry)'
        CALL GC_Error( ErrMsg, RC, ThisLoc )
        RETURN
+    ENDIF
+
+    ! Start "All chem" timer again
+    IF ( Input_Opt%useTimers ) THEN
+       CALL Timer_Start( "All chemistry", RC )
     ENDIF
 
     !========================================================================
@@ -419,6 +432,13 @@ CONTAINS
     !=======================================================================
     ! Convert species units back to original unit
     !=======================================================================
+
+    ! Halt "all chem" timer (so that unit conv can be timed separately)
+    IF ( Input_Opt%useTimers ) THEN
+       CALL Timer_End( "All chemistry", RC )
+    ENDIF
+
+    ! Convert units
     CALL Convert_Spc_Units(                                                  &
          Input_Opt  = Input_Opt,                                             &
          State_Chm  = State_Chm,                                             &
@@ -432,6 +452,11 @@ CONTAINS
        ErrMsg = 'Unit conversion error'
        CALL GC_Error( ErrMsg, RC, ThisLoc )
        RETURN
+    ENDIF
+
+    ! Start "all chem" timer again
+    IF ( Input_Opt%useTimers ) THEN
+       CALL Timer_Start( "All chemistry", RC )
     ENDIF
 
     ! Reset after the first time